Remove static libloudnessCommon
Loudness Common is now a header only library.
