+++ app/SConscript
@@ -1,6 +1,5 @@
 import os
 Import( 'env' )
-Import( 'loudnessCommonLibStatic' )
 Import( 'loudnessAnalyserLibStatic' )
 Import( 'loudnessCorrectorLibStatic' )
 Import( '*' )
@@ -18,7 +17,6 @@ if 'loudnessToolsLibStatic' in locals():
 		'loudness-analyser',
 		Glob( 'analyser/*.cpp' ),
 		LIBS = [
-			loudnessCommonLibStatic,
 			loudnessAnalyserLibStatic,
 			loudnessCorrectorLibStatic,
 			loudnessToolsLibStatic,
@@ -32,7 +30,6 @@ if 'loudnessToolsLibStatic' in locals():
 		'loudness-corrector',
 		Glob( 'corrector/*.cpp' ),
 		LIBS = [
-			loudnessCommonLibStatic,
 			loudnessAnalyserLibStatic,
 			loudnessCorrectorLibStatic,
 			loudnessToolsLibStatic,
@@ -61,7 +58,6 @@ if 'loudnessToolsLibStatic' in locals():
 	    qtEnv.Tool('qt')
 
 	    loudnessValidatorLibraries = [
-	        loudnessCommonLibStatic,
 	        loudnessAnalyserLibStatic,
 	        loudnessCorrectorLibStatic,
 	        loudnessToolsLibStatic,
